✨ highlightText shared/utils로 이동
